Move helper method `viewToSubmission` from support_config to `slack_jira_helpers`
Why:
to be reused also in product config
