+++ test/lib/support_config.test.ts
@@ -1,7 +1,6 @@
 import { merge } from '../helpers';
 import supportConfig from '../../src/lib/support_config';
 import feature from '../../src/util/feature';
-import { ViewSubmission } from '../../src/lib/slack/api_interfaces';
 
 describe('supportConfig', () => {
     const slack_user = { id: 'foo-user-id', name: 'Joe Doe' };
@@ -10,175 +9,6 @@ describe('supportConfig', () => {
         const config = supportConfig('default');
         it('exists', () => {
             expect(config).toBeDefined();
-        });
-
-        describe('#viewToSubmission(view, request_type)', () => {
-            const viewSubmissionView: Partial<ViewSubmission['view']> = {
-                state: {
-                    values: {}
-                }
-            };
-
-            it('returns empty submission object', () => {
-                const submission = config
-                    .viewToSubmission(
-                        viewSubmissionView as ViewSubmission['view'], 'data'
-                    );
-
-                expect(submission).toEqual({});
-            });
-
-            describe('with plain text input', () => {
-                const viewSubmissionView: Partial<ViewSubmission['view']> = {
-                    state: {
-                        values: {
-                            sl_single_value_block: {
-                                sl_single_value: {
-                                    type: 'plain_text_input',
-                                    value: 'Test A'
-                                },
-                            }
-                        }
-                    }
-                };
-
-                it('returns submission with key and value from input', () => {
-                    const submission = config
-                        .viewToSubmission(
-                            viewSubmissionView as ViewSubmission['view'], 'data'
-                        );
-
-                    expect(submission).toEqual({ single_value: 'Test A' });
-                });
-            });
-
-            describe('with static select', () => {
-                const viewSubmissionView: Partial<ViewSubmission['view']> = {
-                    state: {
-                        values: {
-                            ss_region_block: {
-                                ss_region: {
-                                    type: 'static_select',
-                                    selected_option: {
-                                        text: { text: 'Test J' }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                };
-
-                it('returns submission with key and value from select', () => {
-                    const submission = config
-                        .viewToSubmission(
-                            viewSubmissionView as ViewSubmission['view'], 'data'
-                        );
-
-                    expect(submission).toEqual({ region: 'Test J' });
-                });
-
-                describe('when no value selected', () => {
-                    const viewSubmissionView: Partial<ViewSubmission['view']> = {
-                        state: {
-                            values: {
-                                ss_region_block: {
-                                    ss_region: {
-                                        type: 'static_select',
-                                        selected_option: null
-                                    }
-                                }
-                            }
-                        }
-                    };
-
-                    it('returns submission with key and `undefined` as value from select', () => {
-                        const submission = config
-                            .viewToSubmission(
-                                viewSubmissionView as ViewSubmission['view'], 'data'
-                            );
-
-                        expect(submission).toEqual({ region: undefined });
-                    });
-                });
-            });
-
-            describe('with multi select', () => {
-                const viewSubmissionView: Partial<ViewSubmission['view']> = {
-                    state: {
-                        values: {
-                            ms_component_block: {
-                                ms_component: {
-                                    type: 'multi_static_select',
-                                    selected_options: [
-                                        {
-                                            text: { text: 'Test C1' }
-                                        },
-                                        {
-                                            text: { text: 'Test C2' }
-                                        }
-                                    ]
-                                }
-                            }
-                        }
-                    }
-                };
-
-                it('returns submission with key and list of values from select', () => {
-                    const submission = config
-                        .viewToSubmission(
-                            viewSubmissionView as ViewSubmission['view'], 'data'
-                        );
-
-                    expect(submission).toEqual({ component: ['Test C1', 'Test C2'] });
-                });
-
-                describe('when no value selected', () => {
-                    const viewSubmissionView: Partial<ViewSubmission['view']> = {
-                        state: {
-                            values: {
-                                ms_component_block: {
-                                    ms_component: {
-                                        type: 'multi_static_select',
-                                        selected_options: []
-                                    }
-                                }
-                            }
-                        }
-                    };
-
-                    it('returns submission with key and empty array as value from select', () => {
-                        const submission = config
-                            .viewToSubmission(
-                                viewSubmissionView as ViewSubmission['view'], 'data'
-                            );
-
-                        expect(submission).toEqual({ component: [] });
-                    });
-                });
-            });
-
-            describe('with unknown inputtype', () => {
-                const viewSubmissionView: Partial<ViewSubmission['view']> = {
-                    state: {
-                        values: {
-                            sl_unknown_block: {
-                                sl_unknown: {
-                                    type: 'unknown_input_type',
-                                    value: 'Test D'
-                                }
-                            }
-                        }
-                    }
-                };
-
-                it('raise error', () => {
-                    expect(() => {
-                        config.viewToSubmission(
-                            viewSubmissionView as ViewSubmission['view'], 'data'
-                        );
-                    }).toThrowError();
-                });
-            });
         });
     });
